+++ src/constant.h
@@ -472,6 +472,7 @@ constexpr size_t SYS_LIMIT_HIST_SIZE = UINT16_MAX;
 constexpr size_t SYS_LIMIT_ARG_ENTRY_MAX = UINT16_MAX;
 constexpr size_t SYS_LIMIT_ATTR_NUM = 8;
 constexpr size_t SYS_LIMIT_ATTR_CHOICE_SIZE = 16;
+constexpr size_t SYS_LIMIT_XOR_ARG_GROUP_NUM = 63;
 
 constexpr size_t SYS_LINE_RENDERER_TAB_WIDTH = 4;