+++ src/tcerror.h
@@ -237,6 +237,8 @@ DEFINE_TCError(InvalidUDCParamType, "invalid parameter type: `%s', must be deriv
 DEFINE_TCError(InvalidUDCParam, "no-return user-defined command cannot accept parameter");
 DEFINE_TCError(AttrLimit, "number of attributes reaches limit");
 DEFINE_TCError(ChoiceLimit, "number of choice elements reaches limit");
+DEFINE_TCError(XORGroupRange, "xor group number must be 0~63");
+DEFINE_TCError(RequiredXORGroup, "xor group `%d' need `required' param");
 
 DEFINE_TCWarn(MeaninglessCast, "meaningless cast op");
 DEFINE_TCWarn(VarShadowing, "`%s' hides already defined name of outer scope");